rename package constant to images
