+++ internal/images/images.go
@@ -1,7 +1,7 @@
 // This file is generated by "make generate".
 // Don't edit, update .env file and run make target generate.
 
-package constant
+package images
 
 const (
 	DefaultFluentBitExporterImage = "europe-docker.pkg.dev/kyma-project/prod/directory-size-exporter:v20241212-e4adf27f"
